# backend/docs/swagger.yaml
openapi: 3.0.0
servers:
    - description: localhost:5000
      url: http://localhost:5000/api
info:
    description: Plant care system
    version: '1.0.0'
    title: Plan-ts API
    contact:
        email: 'imgjeits@stud.ntnu.no'
    license:
        name: Attribution-NoDerivs 3.0 Unported (CC BY-ND 3.0)
        url: 'https://creativecommons.org/licenses/by-nd/3.0/'
...
paths:
...
    /feedback:
        get:
            summary: List all feedback
            tags:
                - feedback
                - managers
            description: Get a list of all feedback objects in the database
            operationId: listFeedback
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Successful query
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        _id:
                                            $ref: '#/components/schemas/Feedback/properties/_id'
                                        name:
                                            $ref: '#/components/schemas/Feedback/properties/name'
                                        plant_id:
                                            type: object
                                            properties:
                                                _id:
                                                    type: string
                                                    example: '60856ac06bf41a7270f65e62'
                                                    description: 'ObjectID of the plant'
                                                name:
                                                    $ref: '#/components/schemas/Plant/properties/name'
                                        message_body:
                                            $ref: '#/components/schemas/Feedback/properties/message_body'
                                        createdAt:
                                            $ref: '#/components/schemas/Feedback/properties/createdAt'
                                        updatedAt:
                                            $ref: '#/components/schemas/Feedback/properties/updatedAt'
                                        __v:
                                            type: number
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: Feedback not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Could not find any feedback.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to find feedback. [SERVER ERROR MESSAGE]'
        post:
            summary: Add new feedback
            tags:
                - feedback
                - all
            description: Create new feedback and save to database
            operationId: createFeedback
            requestBody:
                description: Info about the feedback
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    $ref: '#/components/schemas/Feedback/properties/name'
                                plant_id:
                                    $ref: '#/components/schemas/Feedback/properties/plant_id'
                                message_body:
                                    $ref: '#/components/schemas/Feedback/properties/message_body'
            responses:
                '201':
                    description: Successfully created new feedback, returns the feedback and a success message.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: 'New feedback successfully saved to database.'
                                    feedback:
                                        $ref: '#/components/schemas/Feedback'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to save the feedback. [SERVER ERROR MESSAGE]'
    /feedback/{id}:
        get:
            summary: Get information about specific feedback
            tags:
                - feedback
                - managers
            description: Get information about specific feedback by its ID.
            operationId: getFeedback
            parameters:
                - name: id
                  in: path
                  required: true
                  description: ObjectID of the feedback
                  schema:
                      type: string
                      example: 6086c0b2e00c2a888c26af05
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Successful query, returns the feedback object.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    _id:
                                        $ref: '#/components/schemas/Feedback/properties/_id'
                                    name:
                                        $ref: '#/components/schemas/Feedback/properties/name'
                                    plant_id:
                                        type: object
                                        properties:
                                            _id:
                                                type: string
                                                example: '60856ac06bf41a7270f65e62'
                                                description: 'ObjectID of the plant'
                                            name:
                                                $ref: '#/components/schemas/Plant/properties/name'
                                    message_body:
                                        $ref: '#/components/schemas/Feedback/properties/message_body'
                                    createdAt:
                                        $ref: '#/components/schemas/Feedback/properties/createdAt'
                                    updatedAt:
                                        $ref: '#/components/schemas/Feedback/properties/updatedAt'
                                    __v:
                                        type: number
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: Could not find the feedback.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'No feedback with this ID.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while looking for feedback with ID 606b7b5993e1ee1504b037e1. [SERVER ERROR MESSAGE]'
        delete:
            summary: Delete feedback
            tags:
                - feedback
                - managers
            description: Delete specific feedback by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  description: ObjectID of the feedback
                  schema:
                      type: string
                      example: 6086c0b2e00c2a888c26af05
            operationId: deleteFeedback
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: 'Successfully deleted feedback.'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: 'Successfully deleted feedback.'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: 'Could not find feedback'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'No feedback with this ID.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to delete feedback. [SERVER ERROR MESSAGE]'
...
components:
    schemas:
...
        Feedback:
            type: object
            description: Feedback sent from users of the app to the managers
            required:
                - name
                - plant_id
                - message_body
            properties:
                _id:
                    type: string
                    description: 'Unique identifier. ObjectID automatically generated by MongoDB'
                    example: '60856ac06bf41a7270f65e62'
                name:
                    type: string
                    description: 'Name of person who sends the feedback'
                    example: 'Michael'
                plant_id:
                    type: string
                    description: ObjectID of the [plant](1.0.0#/Plant) that the feedback is about
                    example: '606b7b5993e1ee1504b037e1'
                message_body:
                    type: string
                    description: 'The message from the user'
                    example: 'This plant looks like it is dying'
                createdAt:
                    type: string
                    format: date-time
                    description: 'Date and time the object was created. Automatically generated.'
                updatedAt:
                    type: string
                    format: date-time
                    description: 'Date and time the object was last updated. Automatically generated.'
                __v:
                    type: number
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    responses:
        UnauthorizedError:
            description: Access token is missing or invalid